Refactoring Code
