+++ CardGame/CardGame/InputView.swift
@@ -7,10 +7,42 @@
 //
 
 import Foundation
-//        카드 게임 종류를 선택하세요.
-//        참여할 사람의 인원을 입력하세요.
+
+enum GameType : String {
+    case seven = "1"
+    case five = "2"
+    
+    var number: Int {
+        switch self {
+        case .seven:
+            return 7
+        case .five:
+            return 5
+        }
+    }
+}
+
+enum NumberOfPlayers : String {
+    case one = "1"
+    case two = "2"
+    case three = "3"
+    case four = "4"
+    
+    var number: Int {
+        switch self {
+        case .one:
+            return 1
+        case .two:
+            return 2
+        case .three:
+            return 3
+        case .four:
+            return 4
+        }
+    }
+}
+
 struct InputView {
-    // 입력값 받는 함수
     public static func readGameType() -> String? {
         let message =
         """
